+++ CHANGES.md
@@ -1,8 +1,9 @@
 0.2.0 (2018-08-08)
 ------------------
+- End-to-end tests for all, and a unit test for updating traces
 - Updating traces on science frames
-- Master trace updating
-- Master trace creation
+- Updating master traces on new lampflats
+- Blind master traces from a single lampflat
 - Master bias and dark creation, subtraction, normalization
 
 0.1.0 (2018-07-13)