+++ CHANGES.md
@@ -1,3 +1,7 @@
+0.2.0 (2018-10-30)
+------------------
+- Updated how the console entry points are called based on upstream updates.
+
 0.1.0 (2018-07-13)
 ------------------
 - Initial setup of the package infrastructure